Don’t use gtk_draw_insertion_cursor()
It makes text caret stem width relative to its height which varies
greatly in our case from an item to another and often gives very thin or
very thick lines. Replaced it with a (simplified) implementation that
uses font size instead, so things are consistent.

+++ src/widget/gtkmathview_common.h
@@ -38,9 +38,9 @@ extern "C" {
   typedef GdkPoint GtkMathViewPoint;
 
   typedef struct _GtkMathViewBoundingBox {
-    gint width;
-    gint height;
-    gint depth;
+    double width;
+    double height;
+    double depth;
   } GtkMathViewBoundingBox;
 
   typedef struct _GtkMathView       GtkMathView;
